requirements.txt
streamlit==1.32.0
groq==0.9.0
httpx==0.27.2
faiss-cpu==1.7.4
sentence-transformers==2.7.0
pymupdf==1.24.0
python-pptx==0.6.23
beautifulsoup4==4.12.3
requests==2.31.0
edge-tts==6.1.9
pydub==0.25.1
numpy==1.26.4
python-dotenv==1.0.1
torch==2.2.2
transformers==4.40.0
tiktoken==0.6.0